Remove workarounds for ISPN-10248.
