+++ clustering/infinispan/client/src/main/java/org/wildfly/clustering/infinispan/client/manager/RemoteCacheManager.java
@@ -26,7 +26,6 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Function;
 
-import org.infinispan.client.hotrod.RemoteCache;
 import org.infinispan.client.hotrod.configuration.Configuration;
 import org.infinispan.client.hotrod.configuration.NearCacheConfiguration;
 import org.infinispan.client.hotrod.event.impl.ClientListenerNotifier;
@@ -39,8 +38,6 @@ import org.wildfly.clustering.infinispan.client.RemoteCacheContainer;
  * @author Paul Ferraro
  */
 public class RemoteCacheManager extends org.infinispan.client.hotrod.RemoteCacheManager implements RemoteCacheContainer {
-    // Workaround for ISPN-10248, used to capture cache name for use by createNearCacheService(...)
-    private static final ThreadLocal<String> CURRENT_CACHE_NAME = new ThreadLocal<>();
 
     private final Map<String, Function<ClientListenerNotifier, NearCacheService<?, ?>>> nearCacheFactories = new ConcurrentHashMap<>();
     private final String name;
@@ -53,17 +50,6 @@ public class RemoteCacheManager extends org.infinispan.client.hotrod.RemoteCache
     @Override
     public String getName() {
         return this.name;
-    }
-
-    @Override
-    public <K, V> RemoteCache<K, V> getCache(String cacheName) {
-        // Workaround for ISPN-10248, capture cache name for use by createNearCacheService(...)
-        CURRENT_CACHE_NAME.set(cacheName);
-        try {
-            return super.getCache(cacheName);
-        } finally {
-            CURRENT_CACHE_NAME.remove();
-        }
     }
 
     @SuppressWarnings("unchecked")
@@ -80,10 +66,9 @@ public class RemoteCacheManager extends org.infinispan.client.hotrod.RemoteCache
     }
 
     @Override
-    protected <K, V> NearCacheService<K, V> createNearCacheService(NearCacheConfiguration config) {
-        String cacheName = CURRENT_CACHE_NAME.get();
+    protected <K, V> NearCacheService<K, V> createNearCacheService(String cacheName, NearCacheConfiguration config) {
         @SuppressWarnings("unchecked")
         Function<ClientListenerNotifier, NearCacheService<K, V>> factory = (cacheName != null) ? (Function<ClientListenerNotifier, NearCacheService<K, V>>) (Function<?, ?>) this.nearCacheFactories.get(cacheName) : null;
-        return (factory != null) ? factory.apply(this.listenerNotifier) : super.createNearCacheService(config);
+        return (factory != null) ? factory.apply(this.listenerNotifier) : super.createNearCacheService(cacheName, config);
     }
 }
